mv and rename subpackage data
